requirements.txt
pytest==6.2.5
pytest-cov==3.0.0
sphinx==4.3.0
black==22.3.0
jupyterlab==3.2.4
pandas==1.3.4
scikit-learn==1.0.1
xgboost==1.5.0

dash==2.3.1
flask==2.0.2
msal==1.16.0
plotly==5.4.0
PyYAML==6.0
# TODO: remove when we migrate from flask-session, see https://bainco.atlassian.net/browse/AE-192
flask-session==0.4.0

